docs: examples
